API

QuantumHamiltonianParticle.ParticleHilbertSpace — Type

ParticleHilbertSpace{PS, BR, QN} <: AbstractHilbertSpace{QN}

Particle Hilbert space.

Example

ParticleHilbertSpace([site1, site2, site3, site4])

QuantumHamiltonianParticle.ParticleSite — Type

ParticleSite{PS<:ParticleSector, BR<:Unsigned, QN<:Tuple{Vararg{<:AbstractQuantumNumber}}}

Particle site.

Example

ParticleSite([state1, state2, state3, state4])

QuantumHamiltonianParticle.ParticleState — Type

ParticleState{PS, BR, QN}

A state, represented by particle occupation.

QuantumHamiltonianParticle.ParticleState — Method

ParticleState(::Type{PS}, name, occvec, ::Type{BR}=UInt)

Create a particle state with no quantum number.

Arguments

PS: particle sector

name: name of the state

occvec::AbstractVector{<:Integer}: occupation Vector

BR: binary type

QuantumHamiltonianParticle.ParticleState — Method

ParticleState(::Type{PS}, name, occvec, quantum_number, ::Type{BR}=UInt)

Create a particle state

Arguments

PS: particle sector

name: name of the state

occvec::AbstractVector{<:Integer}: occupation Vector

quantum_number: quantum number (tuple or integer)

BR: binary type

QuantumHamiltonianParticle.Spin — Type

Spin{Species, N}

Spin type. S⁺ is annihilation, and S⁻ is creation.

QuantumHamiltonian.bitoffset — Method

bitoffset(phs, iptl, isite)

Get the bit offset of the particle iptl at site isite.

QuantumHamiltonian.bitwidth — Method

bitwidth(phs)

Return number of bits needed to represent basis states of phs.

QuantumHamiltonian.compress — Method

compress(hs, indexarray, [type])

Return the binary representation of the basis state represented by indexarray, optionally in type type.

Arguments

hs::ParticleHilbertSpace{PS, BR, QN}

indexarray::CartesianIndex

type::Type{BR2}=BR

QuantumHamiltonian.extract — Method

extract(hs, occbin)

Return the CartesianIndex representation of the basis state represented by occbin.

Arguments

hs::ParticleHilbertSpace

occbin::Unsigned

QuantumHamiltonian.get_bitmask — Method

get_bitmask(phs, [iptl, isite])

Get the bit mask for the particles iptl at sites isite. iptl or isite can either be integer, a vector of integers, or colon :. Bitwise or is taken over list of iptl.

QuantumHamiltonian.get_bitmask — Method

get_bitmask(phs, iptl, isite)

Get the bit mask for the particles iptl at sites isite. iptl or isite can either be integer, a vector of integers, or colon (:). Bitwise or is taken over list of iptl.

QuantumHamiltonian.get_quantum_number — Method

get_quantum_number(phs, statevec)

Get quantum number of the basis state statevec.

QuantumHamiltonian.get_quantum_number — Method

get_quantum_number(phs, bitrep)

Get quantum number of the basis state bitrep.

QuantumHamiltonian.uncompress — Method

uncompress(hs, occbin)

Return the CartesianIndex representation of the basis state represented by occbin.

Arguments

hs::ParticleHilbertSpace

occbin::Unsigned

QuantumHamiltonianParticle.electron_system — Method

electron_system()

Create an electron particle sector and creation/annihilation operators. Returns (particle_sector, c, cdag).candcdagare functions that take site index and spin, e.g.c(3, :up),cdag(4, :↓)`

QuantumHamiltonianParticle.get_fermion_parity — Method

get_fermion_parity(phs, op, bvec)

Get the fermion parity (0 or 1) for when op is applied to bvec.

QuantumHamiltonianParticle.get_fermion_parity — Method

get_fermion_parity(phs, op, bvec)

Get the fermion parity (0 or 1) for when op is applied to bvec.

QuantumHamiltonianParticle.get_occupancy — Method

get_occupancy(phs, iptl, isite, bvec::Unsigned)

Get occupancy of particle iptl at site isite for the given basis state bvec.

QuantumHamiltonianParticle.get_parity_bitmask — Method

get_parity_bitmask(phs, iptl, isite)

Get parity bitmask (i.e. position of the Wigner-Jordan string). Nonzero only for fermions.

QuantumHamiltonianParticle.locvec2statevec — Method

locvec2statevec

particles : Vector of (Vector of particle location)

Example

  P S 1 2 3 4 5 = |0, e↓, e↑, m, m⟩
  e↑  0 0 1 0 0
  e↓  0 1 0 0 0 = c†(m,4) c†(m,5) c†(e↑,3) c†(e↓,2) |Ω⟩
  m   0 0 0 1 1

QuantumHamiltonianParticle.orbitalpermutation — Method

Example

orbitalpermutation(PS, :px => (:py, 1), :py => (:px, -1))

QuantumHamiltonianParticle.set_occupancy — Method

set_occupancy(phs, iptl, isite, bvec::Unsigned, count)

Set occupancy of particle iptl at site isite for the given basis state bvec to count.

QuantumHamiltonianParticle.set_occupancy — Method

set_occupancy(phs, iptl, isite, bvec::Unsigned, count)

Set occupancy of particle iptl at site isite for the given basis state bvec to count.